requirements.txt
nost-tools==3.0.5
tatc==3.4.8
dash==3.0.0
dash-daq==0.6.0
flask==3.0.3
scipy==1.13.1
matplotlib==3.9.4
xarray==2024.7.0
rioxarray==0.15.0
geopandas==1.0.1
netCDF4==1.7.2
aiobotocore==2.21.1
botocore==1.37.1
boto3==1.37.1
s3fs==2025.3.0
pulp==3.1.1
h5netcdf==1.6.1